Update docs to use `gcloud organizations` when adding roles/iam.organizationRoleViewer role to J1 SA
This commit fixes a typo in the development documentation in the `gcloud` command which adds the `roles/iam.organizationRoleViewer` role to the service account JupiterOne uses.

Per the GCP docs (https://cloud.google.com/iam/docs/understanding-roles#iam.organizationRoleViewer), the lowest-level resource on which that role can be granted is the Organization itself. Attempting to run the command as-is (i.e. with `gcloud projects`) results in an error message:

```
ERROR: Policy modification failed. For a binding with condition, run "gcloud alpha iam policies lint-condition" to identify issues in condition.
ERROR: (gcloud.projects.add-iam-policy-binding) INVALID_ARGUMENT: Role (roles/iam.organizationRoleViewer) does not exist in the resource's hierarchy.
```